+++ CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+ - Expose the Command which led to `InvalidCommandSenderException`s
+ - Expose the CommandContext which led to `CommandExecutionException`s
+
 ## [1.3.0] - 2020-12-18
 
 ### Added